Files for Facedetect
Soo about 86% of this looks like Welsh to me. I've tried to understand
it and think I understand some of what's going on in the big picture of
it. I think I have all the files added that are needed, and in video.py
and common.py, there's a lot of extra code we can get rid of.

Also just added the two haarscascades mentioned in facedetect.py.
There's some other file calls such as "lena" for fallback and presets
that I'm unsure what to change to.

Hopefully we can get this functioning tomorrow with little trouble. :)
See you then.